+++ char_text.py
