+++ edit_server_ui.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Form implementation generated from reading ui file 'servereditor.ui'
+# Form implementation generated from reading ui file 'server_editor.ui'
 #
 # Created by: PyQt5 UI code generator 5.13.0
 #
