+++ lb2_rc.py
